TextClassification developed to stable state
TextClassification has been sufficiently developed to a point where it
can be considered stable. Work needs to be done on the hashing database.
Specific changes and why:

* Removed the verbose flag from the .ci/docker-compose
* Added isort as a required pre-commit to clean up headers
* Added __gt__ and __eq__ slots to DataPoint to enable sorting by confidence score
* Fixed TextClassifier to properly unpack and store results from azure into ClassifiedText
* Updated test cases for both Text Classification elements

As of right now, we're lacking solid test coverage on TextClassification
purely because I'd like to make progress on the project as opposed to
learn the ins-and-outs of API testing. That being said I will return.

Likewise, the Zip unpack feature also lacks testing, mostly because
I don't know how to mock a Zip bomb without creating an actual Zip bomb.